Add files via upload
A signal is a set of data or information. A signal is of two types continuous and discrete. continous vary according to time where as discrete vary in integer numbers. We have plotted 10 different functions on MATLAB in continuous and discrete form for better understanding of signals of different kinds. We have also learnt few common functions for more detailing of graphs.